Move util/runner to runner
