+++ pkg/installer/installation/controlplane.go
@@ -21,9 +21,9 @@ import (
 	"time"
 
 	kubeoneapi "github.com/kubermatic/kubeone/pkg/apis/kubeone"
+	"github.com/kubermatic/kubeone/pkg/runner"
 	"github.com/kubermatic/kubeone/pkg/ssh"
 	"github.com/kubermatic/kubeone/pkg/util/context"
-	"github.com/kubermatic/kubeone/pkg/util/runner"
 )
 
 func joinControlplaneNode(ctx *context.Context) error {
