+++ pkg/installer/installation/kubeadm_leader.go
@@ -20,9 +20,9 @@ import (
 	"strconv"
 
 	kubeoneapi "github.com/kubermatic/kubeone/pkg/apis/kubeone"
+	"github.com/kubermatic/kubeone/pkg/runner"
 	"github.com/kubermatic/kubeone/pkg/ssh"
 	"github.com/kubermatic/kubeone/pkg/util/context"
-	"github.com/kubermatic/kubeone/pkg/util/runner"
 )
 
 const (
