+++ pkg/util/context/task.go
@@ -23,8 +23,8 @@ import (
 	"github.com/pkg/errors"
 
 	kubeoneapi "github.com/kubermatic/kubeone/pkg/apis/kubeone"
+	"github.com/kubermatic/kubeone/pkg/runner"
 	"github.com/kubermatic/kubeone/pkg/ssh"
-	"github.com/kubermatic/kubeone/pkg/util/runner"
 )
 
 // NodeTask is a task that is specifically tailored to run on a single node.
